Make add_openmp_flags_if_available() work for clang
The small openmp C test code used to determine if openmp works, fails
to build with clang due to missing include and library paths for
openmp. When run, it then fails due to rpath (runtime path) issues.

This PR adds the necessary include, library, and runtime paths to the
build.

Signed-off-by: James Noss <jnoss@stsci.edu>